+++ kernel/syscall.c
@@ -316,6 +316,9 @@ struct frame *syscall_dispatch(struct frame *f) {
         case NV_SUB_INPUT:
             result = input_ioctl(b, c);
             break;
+        case NV_SUB_AUDIO:
+            result = audio_ioctl(b, c);
+            break;
         default:
             result = -NV_EINVAL;
             break;